pyproject.toml
[project]
name = "honeyhive-daemon"
version = "0.4.1"
description = "Minimal Claude Code telemetry daemon for HoneyHive"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "Apache-2.0"}
authors = [
    {name = "HoneyHive", email = "support@honeyhive.ai"},
]
dependencies = [
    "click>=8.1.0",
    "httpx>=0.27.0,<1.0",
    "pyyaml>=6.0",
    "honeyhive>=1.0.0rc19",
]

[project.urls]
Homepage = "https://honeyhive.ai"
Repository = "https://github.com/honeyhiveai/honeyhive-daemon"

[project.scripts]
honeyhive-daemon = "honeyhive_daemon.main:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.sdist]
include = [
    "honeyhive_daemon/",
    "evaluators/",
    "devin/",
    "enforcement/",
    "tests/",
    "specs/",
    "pyproject.toml",
    "README.md",
    "LICENSE",
    ".github/",
]

[tool.hatch.build.targets.wheel]
packages = ["honeyhive_daemon"]

[tool.hatch.build.targets.wheel.force-include]
"honeyhive_daemon/mappings/claude_code.yaml" = "honeyhive_daemon/mappings/claude_code.yaml"
